cleaned up linkMap, renamed component for clarity
