+++ src/applications/personalization/profile/components/notification-settings/AddContactInfoLink.jsx
@@ -1,34 +1,46 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import { connect } from 'react-redux';
 
 import { getContactInfoDeepLinkURL } from '@@profile/helpers';
+import { showProfileLGBTQEnhancements } from '@@profile/selectors';
 
-import { FIELD_NAMES } from '@@vap-svc/constants';
+import { FIELD_NAMES, MISSING_CONTACT_INFO } from '@@vap-svc/constants';
 
-export const MISSING_CONTACT_INFO = {
-  EMAIL: 'EMAIL',
-  MOBILE: 'MOBILE',
-};
-
-const linkMap = {
-  [MISSING_CONTACT_INFO.EMAIL]: {
-    linkText: 'Add your email address',
-    linkTarget: getContactInfoDeepLinkURL(FIELD_NAMES.EMAIL, true),
-  },
-  [MISSING_CONTACT_INFO.MOBILE]: {
-    linkText: 'Add your mobile phone number',
-    linkTarget: getContactInfoDeepLinkURL(FIELD_NAMES.MOBILE_PHONE, true),
-  },
-};
-
-const AddContactInfoLink = ({ missingInfo }) => {
+const AddContactInfoLink = ({
+  missingInfo,
+  shouldShowProfileLGBTQEnhancements,
+}) => {
   const linkInfo = React.useMemo(
     () => {
+      const linkMap = {
+        [MISSING_CONTACT_INFO.EMAIL]: {
+          linkText: 'Add your email address',
+          linkTarget: getContactInfoDeepLinkURL(
+            FIELD_NAMES.EMAIL,
+            true,
+            shouldShowProfileLGBTQEnhancements,
+          ),
+        },
+        [MISSING_CONTACT_INFO.MOBILE]: {
+          linkText: 'Add your mobile phone number',
+          linkTarget: getContactInfoDeepLinkURL(
+            FIELD_NAMES.MOBILE_PHONE,
+            true,
+            shouldShowProfileLGBTQEnhancements,
+          ),
+        },
+      };
+
       return linkMap[missingInfo];
     },
-    [missingInfo],
+    [missingInfo, shouldShowProfileLGBTQEnhancements],
   );
   return <Link to={linkInfo.linkTarget}>{linkInfo.linkText}</Link>;
 };
 
-export default AddContactInfoLink;
+const mapStateToProps = state => ({
+  shouldShowProfileLGBTQEnhancements: showProfileLGBTQEnhancements(state),
+});
+
+export default connect(mapStateToProps)(AddContactInfoLink);
