+++ src/applications/personalization/profile/components/notification-settings/MissingContactInfoAlertLink.jsx
@@ -5,43 +5,41 @@ import { connect } from 'react-redux';
 import { showProfileLGBTQEnhancements } from '@@profile/selectors';
 import { getContactInfoDeepLinkURL } from '@@profile/helpers';
 
-import { FIELD_NAMES } from '@@vap-svc/constants';
+import { FIELD_NAMES, MISSING_CONTACT_INFO } from '@@vap-svc/constants';
 
-export const MISSING_CONTACT_INFO = {
-  ALL: 'ALL',
-  EMAIL: 'EMAIL',
-  MOBILE: 'MOBILE',
-};
-
-const makeLinkMap = (missingInfo, flag = false) => {
-  const linkMap = {
-    [MISSING_CONTACT_INFO.ALL]: {
-      linkText: 'Update your contact information',
-      linkTarget: getContactInfoDeepLinkURL('phoneNumbers', false, flag),
-    },
-    [MISSING_CONTACT_INFO.EMAIL]: {
-      linkText: 'Add an email address to your profile',
-      linkTarget: getContactInfoDeepLinkURL(FIELD_NAMES.EMAIL, true, flag),
-    },
-    [MISSING_CONTACT_INFO.MOBILE]: {
-      linkText: 'Add a mobile phone number to your profile',
-      linkTarget: getContactInfoDeepLinkURL(
-        FIELD_NAMES.MOBILE_PHONE,
-        true,
-        flag,
-      ),
-    },
-  };
-  return linkMap[missingInfo];
-};
-
-const AddContactInfoLink = ({
+const MissingContactInfoAlertLink = ({
   missingInfo,
   shouldShowProfileLGBTQEnhancements,
 }) => {
   const linkInfo = React.useMemo(
     () => {
-      return makeLinkMap(missingInfo, shouldShowProfileLGBTQEnhancements);
+      const linkMap = {
+        [MISSING_CONTACT_INFO.ALL]: {
+          linkText: 'Update your contact information',
+          linkTarget: getContactInfoDeepLinkURL(
+            'phoneNumbers',
+            false,
+            shouldShowProfileLGBTQEnhancements,
+          ),
+        },
+        [MISSING_CONTACT_INFO.EMAIL]: {
+          linkText: 'Add an email address to your profile',
+          linkTarget: getContactInfoDeepLinkURL(
+            FIELD_NAMES.EMAIL,
+            true,
+            shouldShowProfileLGBTQEnhancements,
+          ),
+        },
+        [MISSING_CONTACT_INFO.MOBILE]: {
+          linkText: 'Add a mobile phone number to your profile',
+          linkTarget: getContactInfoDeepLinkURL(
+            FIELD_NAMES.MOBILE_PHONE,
+            true,
+            shouldShowProfileLGBTQEnhancements,
+          ),
+        },
+      };
+      return linkMap[missingInfo];
     },
     [missingInfo, shouldShowProfileLGBTQEnhancements],
   );
@@ -60,4 +58,4 @@ const mapStateToProps = state => ({
   shouldShowProfileLGBTQEnhancements: showProfileLGBTQEnhancements(state),
 });
 
-export default connect(mapStateToProps)(AddContactInfoLink);
+export default connect(mapStateToProps)(MissingContactInfoAlertLink);
